+++ EDSMResponder/EDSMResponder.cs
@@ -118,7 +118,7 @@ namespace EddiEdsmResponder
                     Dictionary<string, int> data = new Dictionary<string, int>();
                     foreach (MaterialAmount ma in materialInventoryEvent.inventory)
                     {
-                        Material material = Material.FromName(ma.material);
+                        Material material = Material.FromEDName(ma.edname);
                         if (material.category == "Element" || material.category == "Manufactured")
                         {
                             materials.Add(material.EDName, ma.amount);